Encore un peu de réorg
