--- src/definition/index.ts
@@ -1,3 +1,0 @@
-export {Domain} from "./domain";
-export {Entity, EntityField} from "./entity";
-import * as validation from "./validation"; export {validation};
