+++ src/metadata.ts
@@ -17,3 +17,16 @@ export interface Domain {
     DisplayComponent?: ReactType;
     options?: {};
 }
+
+export interface EntityField {
+    domain: Domain;
+    isRequired: boolean;
+    name: string;
+}
+
+export interface Entity<T> {
+    fields: {[name: string]: EntityField};
+    moduleName: string;
+    name: string;
+    type: T;
+}
