+++ tests.tsx
@@ -3,8 +3,8 @@
     Toute modification sera perdue.
 */
 
+import {test, dum} from "./src/testing";
 import * as React from "react";
-import {test, dum} from "./src/testing/base-test";
 
 import {ActionBar} from "./src/search/component/advanced-search/action-bar";
 import {AdvancedSearch} from "./src/search/component/advanced-search";
@@ -13,10 +13,10 @@ import {FacetBox} from "./src/search/component/advanced-search/facet-box";
 import {FacetData} from "./src/search/component/advanced-search/facet-box/facet-data";
 import {GroupComponent} from "./src/search/component/advanced-search/group";
 import {GroupWrapper} from "./src/search/component/group-wrapper";
-import {ListSelection} from "./src/list/component/list-selection";
+import {ListSelection} from "./src/list/list-selection";
 import {ListSummary} from "./src/search/component/advanced-search/list-summary";
-import {ListTable} from "./src/list/component/list-table";
-import {MemoryList} from "./src/list/component/memory-list";
+import {ListTable} from "./src/list/list-table";
+import {MemoryList} from "./src/list/memory-list";
 import {Results} from "./src/search/component/results";
 import {SearchHeader} from "./src/search/component/search-header";
